audit_slide/requirements.txt
python-pptx
colormath
Pillow